move tests of the public API into `tests/` where integration tests live
