+++ examples/sparse_http_reqwest.rs
@@ -1,5 +1,4 @@
-use std::io;
-use crates_index::{Crate, Error, SparseIndex};
+use crates_index::{SparseIndex};
 
 ///
 /// **important**:<br>
